[type: refactor] refactor errors
